Update Dockerfile, AWS templates and README
Updates the README to just the steps required to get the service up and running
through the AWS console. Removed a couple of superfluous security group lines
from the README.

Updated AWS templates to keep naming consistent, and to match what is in the README.

Update the Dockerfile to not copy the code/build inside the image. Instead,
just does a `go get` to get the already compiled binary and run that in our task instead.

As we are using the compiled binary directly, we can than remove the `main.go` file completely.